Create Guess the Number Game
It's a fun game in which system generates a number and asks user guess it under boundary line of 50 counting. System gives multiple attempts in which if guessed number is small system shows "Too low.." and for larger shows "Too High...". At end Code shows output with the number guessed correct message and the number of guesses (attempts by User ).